sletter gammel søknadscontroller og tilhørende tester
